# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PropEdge API - sports betting stats tracker and community
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams
      parameters:
        - name: sport
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a team
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}/stats/{period}:
    get:
      operationId: getTeamStats
      tags: [teams]
      summary: Get team game-by-game stats for a period
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: period
          in: path
          required: true
          schema:
            type: string
            enum: [week, biweek, month, season]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamGameStat"

  # ── Trending / Wow ───────────────────────────────────────────────────────
...
  /bets:
    get:
      operationId: listBets
      tags: [bets]
      summary: List bets for the current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, won, lost, push, void]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bet"
    post:
      operationId: createBet
      tags: [bets]
      summary: Log a new bet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bet"
...
  /bets/{id}:
    patch:
      operationId: updateBet
      tags: [bets]
      summary: Update a bet (settle result)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BetUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bet"
    delete:
      operationId: deleteBet
      tags: [bets]
      summary: Delete a bet
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Simulator ────────────────────────────────────────────────────────────
  /simulator/wallet:
    get:
      operationId: getSimulatorWallet
      tags: [simulator]
      summary: Get virtual wallet balance and stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulatorWallet"
    post:
      operationId: resetSimulatorWallet
      tags: [simulator]
      summary: Reset wallet to a fresh balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletReset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulatorWallet"
  /simulator/bets:
    get:
      operationId: listSimulatorBets
      tags: [simulator]
      summary: List virtual bets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimulatorBet"
    post:
      operationId: createSimulatorBet
      tags: [simulator]
      summary: Place a virtual bet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulatorBetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulatorBet"
  /simulator/bets/{id}:
    patch:
      operationId: settleSimulatorBet
      tags: [simulator]
      summary: Settle a virtual bet (win/loss)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulatorBetSettle"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulatorBet"

  # ── Community ────────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Search users
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
    delete:
      operationId: unfollowUser
      tags: [users]
      summary: Unfollow a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: OK
  /users/{id}/followers:
    get:
      operationId: getUserFollowers
      tags: [users]
      summary: Get a user's followers
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /users/{id}/following:
    get:
      operationId: getUserFollowing
      tags: [users]
      summary: Get users this user follows
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"

  # ── Messages ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Team:
      type: object
      required: [id, name, sport, city, abbreviation]
      properties:
        id:           { type: integer }
        name:         { type: string }
        sport:        { type: string }
        city:         { type: string }
        abbreviation: { type: string }
        logoUrl:      { type: ["string", "null"] }
    TeamGameStat:
      type: object
      required: [id, teamId, gameDate, opponent, isHome]
      properties:
        id:          { type: integer }
        teamId:      { type: integer }
        gameDate:    { type: string }
        opponent:    { type: string }
        isHome:      { type: boolean }
        score:       { type: ["integer", "null"] }
        opponentScore: { type: ["integer", "null"] }
        won:         { type: ["boolean", "null"] }
        # Basketball team stats
        totalPoints:   { type: ["number", "null"] }
        totalRebounds: { type: ["number", "null"] }
        totalAssists:  { type: ["number", "null"] }
        threePointersMade: { type: ["number", "null"] }
        # Baseball team stats
        totalHits:    { type: ["number", "null"] }
        totalRuns:    { type: ["number", "null"] }
        totalHomeRuns: { type: ["number", "null"] }
        errors:       { type: ["number", "null"] }

    # ── Bets ─────────────────────────────────────────────────────────────
    Bet:
      type: object
      required: [id, userId, description, betType, wager, odds, status, createdAt]
      properties:
        id:           { type: integer }
        userId:       { type: integer }
        description:  { type: string }
        betType:      { type: string }
        sportsbook:   { type: ["string", "null"] }
        wager:        { type: number }
        odds:         { type: number }
        potentialPayout: { type: number }
        actualPayout: { type: ["number", "null"] }
        status:       { type: string, enum: [pending, won, lost, push, void] }
        sport:        { type: ["string", "null"] }
        playerName:   { type: ["string", "null"] }
        notes:        { type: ["string", "null"] }
        createdAt:    { type: string }
        settledAt:    { type: ["string", "null"] }
    BetInput:
      type: object
      required: [description, betType, wager, odds]
      properties:
        description:  { type: string }
        betType:      { type: string }
        sportsbook:   { type: string }
        wager:        { type: number }
        odds:         { type: number }
        sport:        { type: string }
        playerName:   { type: string }
        notes:        { type: string }
    BetUpdate:
      type: object
      properties:
        status:       { type: string, enum: [pending, won, lost, push, void] }
        actualPayout: { type: number }
        notes:        { type: string }
...
    SimulatorWallet:
      type: object
      required: [id, userId, balance, startingBalance, totalBets, wins, losses, totalProfit]
      properties:
        id:              { type: integer }
        userId:          { type: integer }
        balance:         { type: number }
        startingBalance: { type: number }
        totalBets:       { type: integer }
        wins:            { type: integer }
        losses:          { type: integer }
        totalProfit:     { type: number }
        winRate:         { type: number }
    WalletReset:
      type: object
      required: [startingBalance]
      properties:
        startingBalance: { type: number }
    SimulatorBet:
      type: object
      required: [id, userId, description, betType, wager, odds, potentialPayout, status, createdAt]
      properties:
        id:              { type: integer }
        userId:          { type: integer }
        description:     { type: string }
        betType:         { type: string }
        wager:           { type: number }
        odds:            { type: number }
        potentialPayout: { type: number }
        actualPayout:    { type: ["number", "null"] }
        status:          { type: string, enum: [pending, won, lost, push] }
        sport:           { type: ["string", "null"] }
        playerName:      { type: ["string", "null"] }
        createdAt:       { type: string }
    SimulatorBetInput:
      type: object
      required: [description, betType, wager, odds]
      properties:
        description: { type: string }
        betType:     { type: string }
        wager:       { type: number }
        odds:        { type: number }
        sport:       { type: string }
        playerName:  { type: string }
    SimulatorBetSettle:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [won, lost, push] }

    # ── Community ────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, username, followersCount, followingCount, createdAt]
      properties:
        id:             { type: integer }
        username:       { type: string }
        displayName:    { type: ["string", "null"] }
        bio:            { type: ["string", "null"] }
        avatarUrl:      { type: ["string", "null"] }
        favoriteSport:  { type: ["string", "null"] }
        followersCount: { type: integer }
        followingCount: { type: integer }
        isFollowing:    { type: boolean }
        createdAt:      { type: string }
    UserProfileUpdate:
      type: object
      properties:
        displayName:   { type: string }
        bio:           { type: string }
        avatarUrl:     { type: string }
        favoriteSport: { type: string }
    FollowResult:
      type: object
      required: [following]
      properties:
        following: { type: boolean }

    # ── Messages ─────────────────────────────────────────────────────────
...