# lib/api-zod/src/generated/types/teamGameStat.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PropEdge API - sports betting stats tracker and community
 * OpenAPI spec version: 0.1.0
 */

export interface TeamGameStat {
  id: number;
  teamId: number;
  gameDate: string;
  opponent: string;
  isHome: boolean;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  opponentScore?: number | null;
  /** @nullable */
  won?: boolean | null;
  /** @nullable */
  totalPoints?: number | null;
  /** @nullable */
  totalRebounds?: number | null;
  /** @nullable */
  totalAssists?: number | null;
  /** @nullable */
  threePointersMade?: number | null;
  /** @nullable */
  totalHits?: number | null;
  /** @nullable */
  totalRuns?: number | null;
  /** @nullable */
  totalHomeRuns?: number | null;
  /** @nullable */
  errors?: number | null;
}